requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-allauth==65.4.1
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
kombu==5.4.2
pillow==11.1.0
prompt_toolkit==3.0.50
python-crontab==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
